+++ CHANGELOG.md
@@ -27,6 +27,7 @@
     but it did not then escape each expanded root path before integrating it into a larger pattern.
   * GUI: In the scan results, some elements could get squished with long file paths.
   * On Windows, some paths were unnecessarily scanned twice.
+  * On Windows, some network share paths were not scanned properly in certain contexts.
 * Changed:
   * GUI: In some cases, Ludusavi would automatically close any open modal
     in order to show a different one,